requirements.txt
asgiref==3.4.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
dataclasses==0.8
Django==3.1.12
django-cors-headers==3.10.1
djangorestframework==3.12.4
djongo==1.3.6
et-xmlfile==1.1.0
firebase-admin==5.2.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gspread==5.1.1
httplib2==0.20.2
idna==3.3
msgpack==1.0.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
pdfkit==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pymongo==3.7.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-http-client==3.3.4
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sendgrid==6.9.4
six==1.16.0
sqlparse==0.2.4
starkbank-ecdsa==2.0.3
twilio==7.4.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
xlrd==1.2.0
